Allow devices to re-integrate cameraserver and mediaserver
Adjust for 7.1

Old (pre-N HALv1) blobs pass virtual addresses of capture buffers to the
encoder. While this was ok before, when the encoders and the camera HAL
lived in the same process, this doesn't work anymore. Fortunately,
it's relatively simple to just go back to the old way of doing things.

To enable:
TARGET_HAS_LEGACY_CAMERA_HAL1 := true
media.stagefright.legacyencoder=true

Change-Id: I4fcc8907ea235b7e83af26122b4da97ca5117816
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>